hotfix optional error, contorller dirction issue
