+++ src/main/kotlin/com/aeyoung/together/domain/study/presentation/controller/StudyRoomController.kt
@@ -2,18 +2,23 @@ package com.aeyoung.together.domain.study.presentation.controller
 
 import com.aeyoung.together.domain.study.StudyRoom
 import com.aeyoung.together.domain.study.presentation.dto.req.CreateStudyRoomReqDto
+import com.aeyoung.together.domain.study.presentation.dto.res.StudyListResDto
+import com.aeyoung.together.domain.study.presentation.dto.res.StudyResDto
 import com.aeyoung.together.domain.study.service.CreateStudyRoomService
+import com.aeyoung.together.domain.study.service.GetAllStudyService
+import com.aeyoung.together.domain.study.service.GetOneStudyService
+import com.aeyoung.together.domain.study.service.SearchStudyService
 import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 
 @RestController
 @RequestMapping("/studies")
 class StudyRoomController(
-    val createStudyRoomService: CreateStudyRoomService
+    private val createStudyRoomService: CreateStudyRoomService,
+    private val getAllStudyService: GetAllStudyService,
+    private val getOneStudyService: GetOneStudyService,
+    private val searchStudyService: SearchStudyService,
 ) {
 
     @PostMapping
@@ -21,4 +26,17 @@ class StudyRoomController(
         createStudyRoomService.createStudyRoom(createStudyRoomReqDto)
         return ResponseEntity.ok().build()
     }
+
+    @GetMapping
+    fun getAllStudy(): ResponseEntity<StudyListResDto> =
+        ResponseEntity.ok(getAllStudyService.execute())
+
+    @GetMapping("/{id}")
+    fun getOneStudy(@PathVariable id: Long): ResponseEntity<StudyResDto> =
+        ResponseEntity.ok(getOneStudyService.execute(id))
+
+    @GetMapping("/filtering")
+    fun searchStudy(@RequestParam searchArgument: String): ResponseEntity<StudyListResDto> =
+        ResponseEntity.ok(searchStudyService.execute(searchArgument))
+
 }
